Move test helpers to Utilities directory.
