+++ src/Containers/JobExplorer/JobExplorer.test.js
@@ -4,7 +4,7 @@ import {
     mountPage,
     preflight200,
     preflight400
-} from '../tests/helpers';
+} from '../../Utilities/tests/helpers';
 import fetchMock from 'fetch-mock-jest';
 import JobExplorer from './JobExplorer';
 import { jobExplorer as constants } from '../../Utilities/constants';
